Organize the client-side logic around actions
This PR organizes the client-side program around first-class
actions. Actions represent what can happen in the system, and
they can have different outcomes. We may want to update the state
of the world responding to certain action, update the UI as a side
effect, or perform an effect (for example AJAX communication) that
may result in further actions.

I've given a explicit type to the state of the application, for now
it's just a type alias. At any given time, the program has a certain
state. Whenever an action occurs, a reducing function with the signature
`(s: State, a: Action): State` yields the new state.

Since the program state changes over time, it is stored in a reactive
variable.

Since the UI is rendered by the server and the client updates it to
reflect the current state, I've thought that we can have a function
with the signature `(s: State, a: Action): IO[Unit]` that will return
an IO action that updates the DOM responding to the different actions.

Effects such as AJAX communication can also be triggered responding to
actions. For example, we respond to the `CompileExercise` action by
performing an AJAX call to the server. Since these effects can have
different outcomes and we want to give feedback to the user, a function
with the `(s: State, a: Action): Future[Option[Action]]` signature
is used to create effects.

When an effect's response contains a `Some` with an action, we interpret
the action's state changes, UI update and further effects. For example, the
`CompileExercise` effect may trigger a `CompilationSucceeded` or
`CompilationFailed` action depending on the response of the server.

Since the UI is not declarative (not derived from program state) and
the server renders it, my approach to program bootstrap has been:

 - Load initial state of the application
 - Reflect the initial state in the UI
 - Listen to user input and convert it into actions
 - Interpret actions
   + Update state
   + Perform effects
   + Update UI

The plumbing is done using a couple reactive variables for the parts
of the program that change over time: the state and the stream of
actions.